requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
coverage==7.6.1
cryptography==42.0.5
Deprecated==1.2.14
Django==5.0.3
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-markdown==0.8.4
django-markdownx==3.0.1
djangorestframework==3.14.0
drf-yasg==1.21.8
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
Markdown==3.5.2
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.2
pillow==10.2.0
psycopg==3.2.2
psycopg2-binary==2.9.9
pycparser==2.21
PyGithub==2.3.0
PyJWT==2.8.0
PyNaCl==1.5.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.31.0
simplejson==3.19.3
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.1
wrapt==1.16.0